refactor: use shared hooks and utils in area
