+++ src/LineDrawer.tsx
@@ -1,5 +1,5 @@
 // Line drawing component extracted from CesiumViewer
-import { useState, useRef, useEffect, useCallback } from 'react'
+import { useState, useRef, useEffect } from 'react'
 import {
   Viewer,
   ScreenSpaceEventHandler,
@@ -10,8 +10,8 @@ import {
   ConstantProperty,
   Cartesian3,
   Entity,
-  HeightReference,
 } from 'cesium'
+import { useDrawingEntities } from './hooks/useDrawingEntities'
 
 interface LineDrawerProps {
   viewer: Viewer | null
@@ -26,103 +26,20 @@ const LineDrawer = ({ viewer }: LineDrawerProps) => {
   const mousePositionRef = useRef<Cartesian3 | null>(null)
   const selectedLineRef = useRef<Entity | null>(null)
   const selectedAnchorRef = useRef<Entity | null>(null)
-  const anchorsRef = useRef<Entity[]>([])
   const [isLineMode, setIsLineMode] = useState(false)
 
-  const highlightLine = (line: Entity) => {
-    if (line.polyline) {
-      line.polyline.material = new ColorMaterialProperty(Color.RED)
-      line.polyline.width = new ConstantProperty(3)
-    }
-  }
+  const {
+    anchorsRef,
+    highlightLine,
+    unhighlightLine,
+    highlightAnchor,
+    unhighlightAnchor,
+    addAnchor,
+    removeLine,
+    removeAnchor,
+  } = useDrawingEntities(viewer)
 
-  const unhighlightLine = (line: Entity) => {
-    if (line.polyline) {
-      line.polyline.material = new ColorMaterialProperty(Color.YELLOW)
-      line.polyline.width = new ConstantProperty(2)
-    }
-  }
-
-  const highlightAnchor = (anchor: Entity) => {
-    if (anchor.point) {
-      anchor.point.color = new ConstantProperty(Color.RED)
-      anchor.point.pixelSize = new ConstantProperty(10)
-    }
-  }
-
-  const unhighlightAnchor = (anchor: Entity) => {
-    if (anchor.point) {
-      anchor.point.color = new ConstantProperty(Color.ORANGE)
-      anchor.point.pixelSize = new ConstantProperty(8)
-    }
-  }
-
-  const removeLine = useCallback(
-    (line: Entity) => {
-    if (!viewer) {
-      return
-    }
-    viewer.entities.remove(line)
-    const lineWithAnchors = line as Entity & { anchors?: [Entity, Entity] }
-    if (lineWithAnchors.anchors) {
-      for (const anchor of lineWithAnchors.anchors) {
-        const a = anchor as Entity & { connectedLines?: Set<Entity> }
-        a.connectedLines?.delete(line)
-        if (!a.connectedLines || a.connectedLines.size === 0) {
-          viewer.entities.remove(a)
-          anchorsRef.current = anchorsRef.current.filter((e) => e !== a)
-          if (selectedAnchorRef.current === a) {
-            selectedAnchorRef.current = null
-          }
-        }
-      }
-    }
-    },
-    [viewer],
-  )
-
-  const removeAnchor = useCallback(
-    (anchor: Entity) => {
-      if (!viewer) {
-        return
-      }
-      viewer.entities.remove(anchor)
-      anchorsRef.current = anchorsRef.current.filter((e) => e !== anchor)
-      const anchorWithLines = anchor as Entity & { connectedLines?: Set<Entity> }
-      if (anchorWithLines.connectedLines) {
-        for (const line of Array.from(anchorWithLines.connectedLines)) {
-          removeLine(line)
-        }
-      }
-    },
-    [removeLine, viewer],
-  )
-
-  const addAnchor = (position: Cartesian3) => {
-    if (!viewer) {
-      return null
-    }
-    for (const existing of anchorsRef.current) {
-      const pos = existing.position?.getValue(viewer.clock.currentTime)
-      if (pos && Cartesian3.distance(pos, position) < 1) {
-        return existing
-      }
-    }
-    const anchor: Entity = viewer.entities.add({
-      position,
-      point: {
-        pixelSize: 8,
-        color: Color.ORANGE,
-        outlineColor: Color.WHITE,
-        outlineWidth: 1,
-        heightReference: HeightReference.CLAMP_TO_GROUND,
-      },
-    })
-    ;(anchor as Entity & { isAnchor: boolean; connectedLines: Set<Entity> }).isAnchor = true
-    ;(anchor as Entity & { isAnchor: boolean; connectedLines: Set<Entity> }).connectedLines = new Set()
-    anchorsRef.current.push(anchor)
-    return anchor
-  }
+  // removeLine, removeAnchor and addAnchor are provided by useDrawingEntities
 
   const startLineMode = () => {
     if (!viewer) {
